+++ Runtime/ILensIdentifiedSectionProvider.cs
@@ -1,0 +1,7 @@
+namespace KostasBan.Lens
+{
+    public interface ILensIdentifiedSectionProvider : ILensSectionProvider
+    {
+        string SectionId { get; }
+    }
+}