+++ Runtime/LensPackageInfo.cs
@@ -3,6 +3,6 @@ namespace KostasBan.Lens
     public static class LensPackageInfo
     {
         public const string PackageName = "com.kostasban.lens";
-        public const string Version = "1.0.1";
+        public const string Version = "1.1.0";
     }
 }