+++ Tests/Runtime/LensRuntimeConsoleTests.cs
@@ -5,6 +5,19 @@ namespace KostasBan.Lens.Tests
 {
     public sealed class LensRuntimeConsoleTests
     {
+        [TearDown]
+        public void TearDown()
+        {
+            var consoles = Object.FindObjectsByType<LensRuntimeConsole>(FindObjectsInactive.Include, FindObjectsSortMode.None);
+            for (var i = 0; i < consoles.Length; i++)
+            {
+                if (consoles[i] != null)
+                {
+                    Object.DestroyImmediate(consoles[i].gameObject);
+                }
+            }
+        }
+
         [Test]
         public void RefreshIntervalIsClampedToNonNegative()
         {
@@ -25,6 +38,41 @@ namespace KostasBan.Lens.Tests
             {
                 Object.DestroyImmediate(gameObject);
             }
+        }
+
+        [Test]
+        public void TryFindExistingReturnsFalseWhenNoConsoleExists()
+        {
+            Assert.IsFalse(LensRuntimeConsole.TryFindExisting(out var console));
+            Assert.IsNull(console);
+        }
+
+        [Test]
+        public void EnsureExistsCreatesConsoleWhenMissing()
+        {
+            var console = LensRuntimeConsole.EnsureExists("Lens Test Console");
+
+            Assert.IsNotNull(console);
+            Assert.AreEqual("Lens Test Console", console.gameObject.name);
+        }
+
+        [Test]
+        public void EnsureExistsReturnsExistingConsole()
+        {
+            var first = LensRuntimeConsole.EnsureExists("First Lens Console");
+            var second = LensRuntimeConsole.EnsureExists("Second Lens Console");
+
+            Assert.AreSame(first, second);
+            Assert.AreEqual("First Lens Console", second.gameObject.name);
+        }
+
+        [Test]
+        public void TryFindExistingReturnsExistingConsole()
+        {
+            var expected = LensRuntimeConsole.EnsureExists();
+
+            Assert.IsTrue(LensRuntimeConsole.TryFindExisting(out var actual));
+            Assert.AreSame(expected, actual);
         }
     }
 }